rename custom models to models
